Remove ThemeProvider
We were using ThemeProvider to provide the Roboto font to our plugin - this is incorrect, especially because ThemeProvider requires a theme to be passed to it, and we were not doing that. This resulted in some crashes. We are now doing the correct approach of installing the Robot font and requiring it at our plugin entry point.